fix bug with table toggle
