+++ src/ducks/HistoricalBalance/AddressTransactions/index.js
@@ -1,18 +1,9 @@
-import React, { useRef, useMemo, useEffect, useState } from 'react'
-import { COLUMNS } from './columns'
+import React, { useState } from 'react'
 import Section from '../Section'
 import { TabType } from '../defaults'
 import Tab from '../../../components/Tab'
-import { useAddressTransactions } from '../hooks'
-import { useDistributions } from '../Address/AssetsDistribution'
-import PagedTable, { buildPageSizes } from '../../_Table/Paged'
-import { getProjectInfo, useProjects } from '../../../stores/projects'
-import styles from './index.module.scss'
-
-const PAGE_SIZES = buildPageSizes([20, 50])
-const DEFAULT_SLUG = 'ethereum'
-
-const getItemKey = ({ trxHash }) => trxHash
+import TopTransactions from './TopTransactions'
+import RecentTransactions from './RecentTransactions'
 
 const Tabs = ({ tabState }) => (
   <>
@@ -22,89 +13,16 @@ const Tabs = ({ tabState }) => (
 )
 
 const AddressTransactions = ({ settings, walletAssets }) => {
-  const pagesItems = useRef({
-    [TabType.LATEST_TRANSACTIONS]: [],
-    [TabType.TOP_TRANSACTIONS]: []
-  }).current
   const tabState = useState(TabType.LATEST_TRANSACTIONS)
   const activeTab = tabState[0]
-  const projects = useProjects()
-  const distributions = useDistributions(walletAssets)
-  const [project, setProject] = useState(null)
-  const [page, setPage] = useState(0)
-  const { transactions, isLoading } = useAddressTransactions(
-    settings,
-    activeTab,
-    page + 1,
-    false,
-    project
-  )
-  const nextTransactions = useAddressTransactions(
-    settings,
-    activeTab,
-    page + 2,
-    isLoading,
-    project
-  ).transactions
-
-  const items = useMemo(
-    () => {
-      pagesItems[activeTab][page] = transactions
-      pagesItems[activeTab][page + 1] = nextTransactions
-      return pagesItems[activeTab].flat()
-    },
-    [transactions, nextTransactions, activeTab]
-  )
-
-  useEffect(
-    () => {
-      setPage(0)
-    },
-    [activeTab]
-  )
-  useEffect(
-    () => {
-      if (activeTab === TabType.LATEST_TRANSACTIONS) {
-        setProject(null)
-      } else {
-        const slug = distributions[0] ? distributions[0].slug : DEFAULT_SLUG
-        setProject(getProjectInfo(projects, slug))
-      }
-    },
-    [distributions, projects]
-  )
-
-  const itemProps = useMemo(() => ({ ...settings, asset: project }), [
-    settings,
-    project
-  ])
 
   return (
     <Section title={<Tabs tabState={tabState} />}>
-      {activeTab === TabType.LATEST_TRANSACTIONS ? (
-        <PagedTable
-          className={styles.table}
-          columns={COLUMNS}
-          pageSizes={PAGE_SIZES}
-          minRows={10}
-          items={items}
-          itemProps={settings}
-          isLoading={isLoading}
-          onPageChange={setPage}
-          getItemKey={getItemKey}
-        />
-      ) : (
-        <PagedTable
-          className={styles.table}
-          columns={COLUMNS}
-          pageSizes={PAGE_SIZES}
-          minRows={10}
-          items={items}
-          itemProps={itemProps}
-          isLoading={isLoading}
-          onPageChange={setPage}
-          getItemKey={getItemKey}
-        />
+      {activeTab === TabType.LATEST_TRANSACTIONS && (
+        <RecentTransactions settings={settings} />
+      )}
+      {activeTab === TabType.TOP_TRANSACTIONS && (
+        <TopTransactions settings={settings} walletAssets={walletAssets} />
       )}
     </Section>
   )
